requirements.txt
asgiref==3.3.1
boto3==1.16.36
botocore==1.19.36
Django==3.1.4
django-crispy-forms==1.10.0
django-storages==1.10.1
gunicorn==20.0.4
jmespath==0.10.0
Pillow==8.0.1
python-dateutil==2.8.1
pytz==2020.4
s3transfer==0.3.3
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
